templates are loaded and rendered once at build time
